refactored user_profile router and saga to be kept in profile router and saga. also added dummy data and several corrections to database sql code
