+++ src/redux/reducers/profile.reducer.js
@@ -2,6 +2,8 @@ const profileReducer = (state = [], action) => {
     switch(action.type) {
         case 'SET_USER_PROFILE':
             return action.payload;
+        case 'ADD_USER_PROFILE':
+            return action.payload;
         default:
             return state;
     }
